refactor: ♻️ further refactoring of the code base
separate models from ui, services, and main
