+++ src/models/yaml.ts
@@ -1,0 +1,5 @@
+export interface YamlStringState {
+  hasSingleQuotes: boolean;
+  hasDoubleQuotes: boolean;
+  isValueEscapedAlready: boolean;
+}
