+++ tsconfig.json
@@ -17,7 +17,5 @@
     ]
   },
   "include": [
-    "src/*.ts",
-    "test-data/*.ts"
-  ]
+    "src/**/*.ts"]
 }
